+++ src/main/java/com/kh/healthgate/employee/model/dao/EmployeeDao.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import com.kh.healthgate.employee.model.vo.Employee;
 
 
-public interface EmployeeDao extends JpaRepository<Employee, Integer>, JpaSpecificationExecutor<Employee> {
+public interface EmployeeDao extends JpaRepository<Employee, Long>, JpaSpecificationExecutor<Employee> {
 
     Page<Employee> findByStatus(String status, Pageable pageable);
